services/agent_runtime/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenticos-agent-runtime"
version = "0.1.0"
description = "Executes agent workflow graphs (LangGraph)."
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
dependencies = [
    "agenticos-shared",
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "httpx>=0.27,<1",
    "nats-py>=2.7,<3",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "respx>=0.21,<1",
]

[tool.setuptools.packages.find]
where = ["src"]
